last mod

+++ asm/valid_indir.c
@@ -30,7 +30,7 @@ int		valid_indir(char *str, int *tab, int *j, t_prog **prog)
 		i++;
 	if (str[i] != '\0')
 		return (my_erormanager(str + i, ft_tab(tab[0], *j, 3), 2));
-	(*prog)->value[tab[1]] = (unsigned int)ft_atoi(str);
+	(*prog)->value[tab[1]] = ft_atoi(str);
 	return (T_IND);
 }
